LAB 02/requirements.txt
Flask
llama-cpp-python
